Add files via upload
Alpha version of algorithm of the shortest path search, also called the travelling salesman problem. It Works with an arbitrary number of points, but if all of them are on the one side from the point of departure. Probably the method is not the most optimal. Improvement is in plan.